requirements.txt
requests == 2.31.0
toml == 0.10.2
websocket-client~=1.6.4